--- src/pages/NotFound.js
@@ -1,15 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-
-class NotFound extends React.Component {
-  render() {
-    return (
-      <div className="NotFound">
-        <p>Page not found for {this.props.location.pathname}</p>
-        <Link to="/">Go Home</Link>
-      </div>
-    );
-  }
-}
-
-export default NotFound;
