+++ src/pages/about/About.css
@@ -1,0 +1,7 @@
+.About {
+    background-image: url("../../components/assets/images/backgrounds/Outer_space.jpg");
+    height: 100vh;
+    width: 100vw;
+    color: white;
+}
+
